Create Get-SPNs.ps1
Enumerate all SPNs and exports to a CSV file. The CSV file will have two columns:

SPN = the SPNs of the user we've found
WhenChanged = The date of when the password was last changed

This script works in conjunction with LiteRoast.ps1 so you can pull all service principal names and choose which ones to target based on the password changed date.